fix: ValidateProviderAddress to accept non-legacy addresses
Previously ValidateProviderAddress would accept *only* legacy address
which was never the intention during the introduction of the function.

While this would technically be a breaking change, a brief search of
publicly available GitHub code shows no current use of that function.

Also, the original issue requesting the function intended to do
the exact opposite, i.e. ensure the address is *not* legacy.
https://github.com/hashicorp/terraform-registry-address/issues/10